Move bootstap kubeconfig validation func to bootstrap pkg
