+++ pkg/controller/importconfig/cluster_info.go
@@ -6,7 +6,6 @@ package importconfig
 import (
 	"context"
 	"fmt"
-	"reflect"
 	"time"
 
 	klusterletconfigv1alpha1 "github.com/stolostron/cluster-lifecycle-api/klusterletconfig/v1alpha1"
@@ -17,9 +16,7 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/clientcmd"
-	certutil "k8s.io/client-go/util/cert"
 	"k8s.io/klog/v2"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // getBootstrapKubeConfigDataFromImportSecret aims to reuse the bootstrap kubeconfig data if possible.
@@ -45,43 +42,12 @@ func getBootstrapKubeConfigDataFromImportSecret(ctx context.Context, clientHolde
 		return nil, nil, fmt.Errorf("failed to parse kubeconfig data: %v", err)
 	}
 
-	// check if the kube apiserver address is changed
-	validKubeAPIServer, err := validateKubeAPIServerAddress(ctx, kubeAPIServer, klusterletConfig, clientHolder)
+	valid, err := bootstrap.ValidateBootstrapKubeconfig(ctx, clientHolder, klusterletConfig, clusterName,
+		kubeAPIServer, caData, proxyURL, ctxClusterName)
 	if err != nil {
-		return nil, nil, fmt.Errorf("failed to validate kube apiserver address: %v", err)
+		return nil, nil, err
 	}
-	if !validKubeAPIServer {
-		klog.Infof("KubeAPIServer invalid for the managed cluster %s, kubeAPIServer: %v", clusterName, kubeAPIServer)
-		return nil, nil, nil
-	}
-
-	// check if the CA data is changed
-	validCAData, err := validateCAData(ctx, caData, kubeAPIServer, klusterletConfig, clientHolder, clusterName)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to validate CA data: %v", err)
-	}
-	if !validCAData {
-		klog.Infof("CAdata is invalid for the managed cluster %s", clusterName)
-		return nil, nil, nil
-	}
-
-	// check if the proxy url changed
-	validProxyConfig, err := validateProxyConfig(proxyURL, caData, klusterletConfig)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to validate proxy config: %v", err)
-	}
-	if !validProxyConfig {
-		klog.Infof("Proxy config is invalid for the managed cluster %s", clusterName)
-		return nil, nil, nil
-	}
-
-	// check if the current context cluster name of the bootstrap kubeconfig is changed
-	validCtxClusterName, err := validateContextClusterName(ctx, clientHolder.RuntimeClient, ctxClusterName)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to validate context cluster name: %v", err)
-	}
-	if !validCtxClusterName {
-		klog.Infof("Context cluster name is invalid for the managed cluster %s", clusterName)
+	if !valid {
 		return nil, nil, nil
 	}
 
@@ -146,37 +112,6 @@ func parseKubeConfigData(kubeConfigData []byte) (
 	return
 }
 
-func validateKubeAPIServerAddress(ctx context.Context, kubeAPIServer string,
-	klusterletConfig *klusterletconfigv1alpha1.KlusterletConfig,
-	clientHolder *helpers.ClientHolder) (bool, error) {
-	if len(kubeAPIServer) == 0 {
-		return false, nil
-	}
-
-	currentKubeAPIServer, err := bootstrap.GetKubeAPIServerAddress(ctx, clientHolder.RuntimeClient, klusterletConfig)
-	if err != nil {
-		return false, err
-	}
-
-	return kubeAPIServer == currentKubeAPIServer, nil
-}
-
-func validateCAData(ctx context.Context, caData []byte, kubeAPIServer string,
-	klusterletConfig *klusterletconfigv1alpha1.KlusterletConfig,
-	clientHolder *helpers.ClientHolder, clusterName string) (bool, error) {
-	if len(caData) == 0 {
-		// CA data is empty
-		return false, nil
-	}
-
-	currentCAData, err := bootstrap.GetBootstrapCAData(ctx, clientHolder, kubeAPIServer, clusterName, klusterletConfig)
-	if err != nil {
-		return false, err
-	}
-
-	return reflect.DeepEqual(caData, currentCAData), nil
-}
-
 func validateToken(token string, expiration []byte) bool {
 	if len(token) == 0 {
 		// no token in the kubeconfig
@@ -197,55 +132,4 @@ func validateToken(token string, expiration []byte) bool {
 	refreshThreshold := 8640 * time.Hour / 5
 	lifetime := expirationTime.Sub(now.Time)
 	return lifetime > refreshThreshold
-}
-
-func validateProxyConfig(kubeconfigProxyURL string, kubeconfigCAData []byte, klusterletConfig *klusterletconfigv1alpha1.KlusterletConfig) (bool, error) {
-	proxyURL, proxyCAData := bootstrap.GetProxySettings(klusterletConfig)
-	if proxyURL != kubeconfigProxyURL {
-		return false, nil
-	}
-
-	return hasCertificates(kubeconfigCAData, proxyCAData)
-}
-
-func validateContextClusterName(ctx context.Context, client client.Client, clusterName string) (bool, error) {
-	ctxClusterName, err := bootstrap.GetKubeconfigClusterName(ctx, client)
-	if err != nil {
-		return false, err
-
-	}
-	return ctxClusterName == clusterName, nil
-}
-
-// hasCertificates returns true if the supersetCertData contains all the certs in subsetCertData
-func hasCertificates(supersetCertData, subsetCertData []byte) (bool, error) {
-	if len(subsetCertData) == 0 {
-		return true, nil
-	}
-
-	if len(supersetCertData) == 0 {
-		return false, nil
-	}
-
-	superset, err := certutil.ParseCertsPEM(supersetCertData)
-	if err != nil {
-		return false, err
-	}
-	subset, err := certutil.ParseCertsPEM(subsetCertData)
-	if err != nil {
-		return false, err
-	}
-	for _, sub := range subset {
-		found := false
-		for _, super := range superset {
-			if reflect.DeepEqual(sub.Raw, super.Raw) {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false, nil
-		}
-	}
-	return true, nil
 }
